Refactor handicap calculators to return integer results instead of float. Adjust tests accordingly.

+++ src/Calculator/HandicapCalculatorInterface.php
@@ -13,5 +13,5 @@ interface HandicapCalculatorInterface
      * Calculate the playing handicap based on the actual handicap,
      * course slope, course rating, and course par.
      */
-    public function getHandicap(array $options): float;
+    public function getHandicap(array $options): int;
 }
